+++ src/preferences/ShortcutsPage.ts
@@ -43,6 +43,22 @@ export class ShortcutsPage {
             settings: this._settings,
             window: this.window,
             group,
+            key: 'move-workspace-left',
+            title: 'Move current workspace left',
+        });
+
+        addKeyboardShortcut({
+            settings: this._settings,
+            window: this.window,
+            group,
+            key: 'move-workspace-right',
+            title: 'Move current workspace right',
+        });
+
+        addKeyboardShortcut({
+            settings: this._settings,
+            window: this.window,
+            group,
             key: 'activate-previous-key',
             title: 'Switch to previous workspace',
         });